# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.JudgeResult:
    properties:
      compile_error:
        description: 编译错误信息
        type: string
      memory_used:
        description: 内存使用(KB)
        type: integer
      runtime_error:
        description: 运行时错误信息
        type: string
      status:
        $ref: '#/definitions/models.SubmitStatus'
      test_results:
        description: 测试用例结果
        items:
          $ref: '#/definitions/models.TestResult'
        type: array
      time_used:
        description: 时间使用(ms)
        type: integer
    type: object
  models.Language:
    enum:
    - c
    - cpp
    - java
    - python
    - go
    type: string
    x-enum-varnames:
    - LangC
    - LangCpp
    - LangJava
    - LangPython
    - LangGo
...
  models.Submit:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: string
      language:
        $ref: '#/definitions/models.Language'
      problem_id:
        type: string
      result:
        $ref: '#/definitions/models.JudgeResult'
      status:
        $ref: '#/definitions/models.SubmitStatus'
      updated_at:
        type: string
      user_id:
        type: string
    required:
    - code
    - language
    - problem_id
    - user_id
    type: object
  models.SubmitCodeRequest:
    properties:
      code:
        type: string
      language:
        type: string
      problem_id:
        type: string
    required:
    - code
    - language
    - problem_id
    type: object
  models.SubmitStatus:
    enum:
    - pending
    - running
    - accepted
    - wrong_answer
    - time_limit
    - memory_limit
    - runtime_error
    - compile_error
    - system_error
    type: string
    x-enum-comments:
      StatusAccepted: 通过
      StatusCompileError: 编译错误
      StatusMemoryLimit: 内存超限
      StatusPending: 等待中
      StatusRunning: 运行中
      StatusRuntimeError: 运行时错误
      StatusSystemError: 系统错误
      StatusTimeLimit: 超时
      StatusWrongAnswer: 答案错误
    x-enum-descriptions:
    - 等待中
    - 运行中
    - 通过
    - 答案错误
    - 超时
    - 内存超限
    - 运行时错误
    - 编译错误
    - 系统错误
    x-enum-varnames:
    - StatusPending
    - StatusRunning
    - StatusAccepted
    - StatusWrongAnswer
    - StatusTimeLimit
    - StatusMemoryLimit
    - StatusRuntimeError
    - StatusCompileError
    - StatusSystemError
...
  models.TestResult:
    properties:
      error:
        description: 错误信息
        type: string
      expected:
        description: 期望输出
        type: string
      is_sample:
        description: 是否为示例用例（用于判断是否返回详细信息）
        type: boolean
      memory_used:
        description: KB
        type: integer
      output:
        description: 仅示例用例返回以下字段（隐藏用例不返回详细输出）
        type: string
      status:
        $ref: '#/definitions/models.SubmitStatus'
      test_case_id:
        type: string
      time_used:
        description: ms
        type: integer
    type: object
...
  models.UpdateUserRequest:
    properties:
      is_active:
        type: boolean
      role:
        type: string
    type: object
...
  models.UserProfile:
    properties:
      avatar:
        type: string
      bio:
        type: string
      class:
        type: string
      created_at:
        type: string
      email:
        type: string
      grade:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      major:
        type: string
      real_name:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      school:
        type: string
      student_id:
        type: string
      username:
        type: string
    type: object
  models.UserRole:
    enum:
    - admin
    - teacher
    - student
    - contest
    type: string
    x-enum-comments:
      RoleAdmin: 系统管理员
      RoleContest: 竞赛发起者
      RoleStudent: 学生
      RoleTeacher: 老师
    x-enum-descriptions:
    - 系统管理员
    - 老师
    - 学生
    - 竞赛发起者
    x-enum-varnames:
    - RoleAdmin
    - RoleTeacher
    - RoleStudent
    - RoleContest
...
host: localhost:8080
info:
  contact:
    email: support@zkcodearena.com
    name: API Support
    url: http://localhost/support
  description: |-
    ZK Code Arena 在线编程平台 REST API 文档
    提供用户管理、题库管理、班级课程管理、代码评测等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://localhost/terms/
  title: ZK Code Arena API
  version: "1.0"
paths:
...
  /statistics/system:
    get:
      consumes:
      - application/json
      description: 管理员查看系统整体统计信息
      produces:
      - application/json
      responses:
        "200":
          description: 系统统计信息
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 获取失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取系统统计（管理员）
      tags:
      - 统计
  /statistics/user:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的提交统计信息
      produces:
      - application/json
      responses:
        "200":
          description: 用户统计信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的用户ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 获取失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取当前用户统计
      tags:
      - 统计
  /statistics/user/{id}:
    get:
      consumes:
      - application/json
      description: 管理员查看指定用户的提交统计信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用户统计信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的用户ID
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 获取失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取指定用户统计（管理员）
      tags:
      - 统计
  /submit:
    get:
      consumes:
      - application/json
      description: 分页获取提交记录（非管理员仅可查看自己的提交）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 题目ID筛选
        in: query
        name: problem_id
        type: string
      - description: 用户ID筛选（管理员可用）
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 提交列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 获取失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取提交列表
      tags:
      - 提交
    post:
      consumes:
      - application/json
      description: 提交代码进行判题
      parameters:
      - description: 提交信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SubmitCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 提交记录
          schema:
            $ref: '#/definitions/models.Submit'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证用户
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 题目不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 提交失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 提交代码
      tags:
      - 提交
  /submit/{id}:
    get:
      consumes:
      - application/json
      description: 根据提交ID获取详细信息（管理员或提交者本人）
      parameters:
      - description: 提交ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 提交详情
          schema:
            $ref: '#/definitions/models.Submit'
        "400":
          description: 无效的提交ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 提交不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取提交详情
      tags:
      - 提交
...
  /user:
    get:
      consumes:
      - application/json
      description: 分页获取用户列表，可按角色筛选（仅管理员）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 用户角色筛选
        enum:
        - admin
        - teacher
        - student
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用户列表
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 获取失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取用户列表（管理员）
      tags:
      - 用户
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: 管理员删除指定用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的用户ID
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 删除失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除用户（管理员）
      tags:
      - 用户
    get:
      consumes:
      - application/json
      description: 根据用户ID获取用户信息（管理员或本人）
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            $ref: '#/definitions/models.UserProfile'
        "400":
          description: 无效的用户ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证用户
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 用户不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取指定用户信息
      tags:
      - 用户
    put:
      consumes:
      - application/json
      description: 管理员更新用户角色和激活状态
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的用户信息
          schema:
            $ref: '#/definitions/models.UserProfile'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 用户不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 更新失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新用户信息（管理员）
      tags:
      - 用户
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"